+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.10.0]
+
+### Added
+- Update `AudioQualitySelectBox`/`VideoQualitySelectBox` entries when the period of a source changes
+
+### Changed
+- Export bundled UI (`bitmovinplayer-ui.js`) as UMD module (instead of global module)
+
+### Fixed
+- Fix `Uncaught TypeError` when `require`ing UI before player
+- Don't write UI into global namespace (`bitmovin.playerui`) when loaded as module with `require`
+
 ## [2.9.0]
 
 ### Added
@@ -234,6 +246,7 @@ Version 2.0 of the UI framework is built for player 7.1. If absolutely necessary
 ## 1.0.0 - 2017-02-03
 - First release
 
+[2.10.0]: https://github.com/bitmovin/bitmovin-player-ui/compare/v2.9.0...v2.10.0
 [2.9.0]: https://github.com/bitmovin/bitmovin-player-ui/compare/v2.8.3...v2.9.0
 [2.8.3]: https://github.com/bitmovin/bitmovin-player-ui/compare/v2.8.2...v2.8.3
 [2.8.2]: https://github.com/bitmovin/bitmovin-player-ui/compare/v2.8.1...v2.8.2